Correct cyclic service implementation/definition (SpringBoot forbid this since last upgrade): extract checkPassword in a separate service.
